+++ app/javascript/src/components/Invoices/common/InvoiceForm/SendInvoice/index.tsx
@@ -15,7 +15,9 @@ import { Toastr } from "StyledComponents";
 
 import invoicesApi from "apis/invoices";
 import { ApiStatus as InvoiceStatus } from "constants/index";
+import { useUserContext } from "context/UserContext";
 
+import Recipient from "./Recipient";
 import {
   isEmailValid,
   emailSubject,
@@ -30,22 +32,6 @@ interface InvoiceEmail {
   recipients: string[];
 }
 
-const Recipient: React.FC<{ email: string; handleClick: any }> = ({
-  email,
-  handleClick,
-}) => (
-  <div className="space-XIcon-2 m-0.5 flex w-fit items-center rounded-full border bg-miru-gray-400 px-2 py-1">
-    <p>{email}</p>
-    {/* <button
-      className="text-miru-black-1000 hover:text-miru-red-400"
-      type="button"
-      onClick={handleClick}
-    >
-      <XIcon size={14} weight="bold" />
-    </button> */}
-  </div>
-);
-
 const SendInvoice: React.FC<any> = ({
   invoice,
   setIsSending,
@@ -54,6 +40,8 @@ const SendInvoice: React.FC<any> = ({
   isSendReminder = false,
   setIsSendReminder,
 }) => {
+  const { user } = useUserContext();
+
   const [status, setStatus] = useState<InvoiceStatus>(InvoiceStatus.IDLE);
   const [invoiceEmail, setInvoiceEmail] = useState<InvoiceEmail>({
     subject: emailSubject(invoice, isSendReminder),
@@ -177,7 +165,9 @@ const SendInvoice: React.FC<any> = ({
                     "flex flex-wrap rounded bg-miru-gray-100 p-1.5",
                     { "h-9": !invoiceEmail.recipients }
                   )}
-                  // onClick={() => input.current.focus()}
+                  onClick={() =>
+                    user.email == "supriya@saeloun.com" && input.current.focus()
+                  }
                 >
                   {invoiceEmail.recipients.map(recipient => (
                     <Recipient
@@ -186,21 +176,23 @@ const SendInvoice: React.FC<any> = ({
                       key={recipient}
                     />
                   ))}
-                  {/* <input
-                    name="to"
-                    ref={input}
-                    style={{ width }}
-                    type="email"
-                    value={newRecipient}
-                    className={cn(
-                      "focus:outline-none mx-1.5 w-fit cursor-text rounded bg-miru-gray-100 py-2",
-                      {
-                        "text-miru-red-400": !isEmailValid(newRecipient),
-                      }
-                    )}
-                    onChange={e => setNewRecipient(e.target.value.trim())}
-                    onKeyDown={handleInput}
-                  /> */}
+                  {user.email == "supriya@saeloun.com" && (
+                    <input
+                      name="to"
+                      ref={input}
+                      style={{ width }}
+                      type="email"
+                      value={newRecipient}
+                      className={cn(
+                        "focus:outline-none mx-1.5 w-fit cursor-text rounded bg-miru-gray-100 py-2",
+                        {
+                          "text-miru-red-400": !isEmailValid(newRecipient),
+                        }
+                      )}
+                      onChange={e => setNewRecipient(e.target.value.trim())}
+                      onKeyDown={handleInput}
+                    />
+                  )}
                 </div>
               </fieldset>
               <fieldset className="field_with_errors flex flex-col">
