+++ app/javascript/src/components/Invoices/popups/SendInvoice/index.tsx
@@ -17,7 +17,9 @@ import invoicesApi from "apis/invoices";
 import { CustomAdvanceInput } from "common/CustomAdvanceInput";
 import { CustomInputText } from "common/CustomInputText";
 import { CustomTextareaAutosize } from "common/CustomTextareaAutosize";
+import Recipient from "components/Invoices/common/InvoiceForm/SendInvoice/Recipient";
 import { ApiStatus as InvoiceStatus } from "constants/index";
+import { useUserContext } from "context/UserContext";
 
 import {
   isEmailValid,
@@ -33,22 +35,6 @@ interface InvoiceEmail {
   recipients: string[];
 }
 
-const Recipient: React.FC<{ email: string; handleClick: any }> = ({
-  email,
-  handleClick,
-}) => (
-  <div className="space-XIcon-2 m-0.5 flex w-fit items-center rounded-full border bg-miru-gray-400 px-2 py-1">
-    <p>{email}</p>
-    {/* <button
-      className="text-miru-black-1000 hover:text-miru-red-400"
-      type="button"
-      onClick={handleClick}
-    >
-      <XIcon size={14} weight="bold" />
-    </button> */}
-  </div>
-);
-
 const SendInvoice: React.FC<any> = ({
   invoice,
   setIsSending,
@@ -57,6 +43,7 @@ const SendInvoice: React.FC<any> = ({
   // eslint-disable-next-line @typescript-eslint/no-empty-function
   setIsSendReminder = _value => {},
 }) => {
+  const { user } = useUserContext();
   const [status, setStatus] = useState<InvoiceStatus>(InvoiceStatus.IDLE);
   const [invoiceEmail, setInvoiceEmail] = useState<InvoiceEmail>({
     subject: emailSubject(invoice, isSendReminder),
@@ -190,24 +177,28 @@ const SendInvoice: React.FC<any> = ({
                       key={recipient}
                     />
                   ))}
-                  {/* <input
-                    name="to"
-                    ref={input}
-                    style={{ width }}
-                    type="email"
-                    value={newRecipient}
-                    className={cn(
-                      "focus:outline-none mx-1.5 w-fit cursor-text",
-                      {
-                        "text-miru-red-400": !isEmailValid(newRecipient),
-                      }
-                    )}
-                    onChange={e => setNewRecipient(e.target.value.trim())}
-                    onKeyDown={handleInput}
-                  /> */}
+                  {user.email == "supriya@saeloun.com" && (
+                    <input
+                      name="to"
+                      ref={input}
+                      style={{ width }}
+                      type="email"
+                      value={newRecipient}
+                      className={cn(
+                        "focus:outline-none mx-1.5 w-fit cursor-text",
+                        {
+                          "text-miru-red-400": !isEmailValid(newRecipient),
+                        }
+                      )}
+                      onChange={e => setNewRecipient(e.target.value.trim())}
+                      onKeyDown={handleInput}
+                    />
+                  )}
                 </div>
               }
-              // onClick={() => input.current.focus()}
+              onClick={() =>
+                user.email == "supriya@saeloun.com" && input.current.focus()
+              }
             />
           </fieldset>
           <fieldset className="field_with_errors flex flex-col">
